Pythonic changes
Tried to make code conform to PEP8, mostly succeeded.  Fixed some variable shadowing issues changed how get_child_id was getting called at the end- it prints stuff in the method so there really isn't a need to try and capture anything from it.  That should probably be cleaned up, maybe I'll get to it if I feel froggy.  I might also add an option for it to dump things straight to a file if I'm feeling sufficiently cute.